(feat) use delta time to compute position
The goal with this change is to allow the user to instantly start moving
 at speed, instead of depending on key-repeat.

- offset/position should be floating-point because, depending on the
frame rate, it will be updated in non-whole amounts.
- use the last frame's duration to determine how much to change the
position by this frame

+++ src/client_prediction/Client.cpp
@@ -30,12 +30,12 @@ void Client::process_server_messages()
     });
 }
 
-int Client::offset() const
+double Client::offset() const
 {
     return _offset;
 }
 
-void Client::offset(int offset)
+void Client::offset(double offset)
 {
     _offset = offset;
 }
