+++ src/client_prediction/Client.hpp
@@ -15,11 +15,11 @@ class Client {
     std::mutex _queue_mutex;
     std::vector<Message> _queue;
 
-    int _offset{0};
+    double _offset{0.0};
 
 public:
-    void offset(int);
-    [[nodiscard]] int offset() const;
+    void offset(double);
+    [[nodiscard]] double offset() const;
 
     void process_server_messages();
     void send(State_message const& msg, std::chrono::milliseconds delay);
