+++ src/client_prediction/State_message.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
 struct State_message {
-    int position;
+    double position;
 };
